swig-py: Manage living decendant batons by editor/parse_fns3 baton.
In APIs using svn_delta_edit_t/svn_repos_parse_fns3_t, they create
decendant batons along with call back functions.  Generally those
batons are alocated from memory pools, so we don't need deallocate
explicitly even if some API causes an error.  However, on SWIG
Python binings, we uses Python objects as a part of batons and
aquire their reference, so we should releace the references of
them if the ancestor editor/parse_fns3 baton is closed or removed.
To do it, we introduce a double linked list of decendant batons
which are not release a reference of Python object in editor/parse_fns3
baton.

* subversion/bindings/swig/python/libsvn_swig_py/swigutil_py.c
  (svn_swig_py_item_baton_t):
    - Add new members to manage decendant batons
  (make_editor_baton): New function.
  (make_baton): Modified for change of svn_siwg_py_item_baton_t.
  (release_baton): New function.
  (close_baton):
    - Modify to accept new svn_swig_py_item_baton_t * type baton.
    - Use release_baton to release the reference of Python baton object.
    - Release the baton even if error occured.
    - Use release_baton to release reference of the Python object of
      the baton.
  (open_root, add_directory, open_directory, add_file, open_file,
   parse_fn3_new_revision_record, parse_fn3_new_node_record,):
    Pass editor baton itself instead of Python object in it to make
    baton.
  (window_handler):
    Release handler object if handler function causes error for window
    other than None.
  (close_file):
    - Release the baton even if error occured.
    - Use release_baton to release reference of the Python object of
      the baton.
  (close_edit_baton): New function.
  (close_edit, abort_edit, svn_swig_py_parse_fns3_destroy):
    Use close_edit_baton to release references of the Python objects in
    the batons.
  (svn_swig_py_make_editor):
    Use make_editor_baton to make edit_baton.
  (svn_swig_py_make_parse_fns3):
    Use make_editor_baton to make parse_baton.